nginx working, nothing else
Nginx dockerfile looks fine. The image is created successfully each
time, and uses supervisord to ensure that ssh is always running also.

The nginx container can be ssh'd into with the password 'codam2020'.

I've been trying to set up the kubernetes dashboard to work, but after
hours of endless googling I'm still stuck trying to access it when using
the LoadBalancer - I'm using the exact same yaml files as recommended,
and even with the same yaml file as somebody else it just does not seem
to expose the port correctly - and connections time out every time.

Maybe it will magically work tomorrow.