Added quite a lot
Grafana and InfluxDB I think are correctly setup, although the
dashboards need cleaning and some more need to be added for Grafana -
but now upon startup Grafana will be linked to the InfluxDB Database,
and all dashboards are already loaded through 'provisioning'.

Telegraf however, is clearly communicating with InfluxDB correctly, but
it seems to be failing to send the correct metrics, and I imagine this
may be due to some sort of failure in setting up a Service Account
correctly, or I am getting the incorrect information from the Kubernetes
API.

I'll work to fix this today - once these things are set up though, I
think I'll be well placed to aim for completion of the project -
wordpress and MySQL will likely include the same concepts as have been
used previously in ft_server - although the Reverse Proxy from Nginx to
phpmyadmin might require a bit of research.