InfluxDB, Telegraf and Grafana at acceptable level
I haven't managed to get communication between telegraf and the
metrics-server completed, but the dashboard now contains information on
cpu usage and memory.

The kube_inventory plugin of telegraf is working perfectly however, and
that should be sufficient. InfluxDB has permanence, and communicates
easily with telegraf. Grafana has no difficulty in communicating with
InfluxDB also.

At this point, the skeleton structure for mysql has been laid down - the
phpmyadmin tables are created upon startup, the container has a
PersistentVolumeClaim, a lot is looking very good.

Now I'm trying to launch Wordpress, and for the setup of Wordpress a
functioning MySQL is essential - so MySQL will be improved at the same
time.

Following successful setup of MySQL and WordPress, then the next steps
will be:
	- Setup PHPmyadmin, which should hopefully be very similar to
	WordPress.
	- Setup the redirection of nginx to wordpress, and the reverse
	proxy to PHPmyadmin. While I've never done this before, I can't
	see it being much more than some small edits to the nginx .conf
	files.
	- Launch FTPS, with successful communication to MySQL. Hopefully
	this won't be too difficult either.

Once all of these steps are completed, then I should clean up a lot of
the config files, add comments where necessary, and finally try to do
some mock evaluations with other people.