FTP at basic functionality
I've managed to get the FTP container up and running, using vsftpd (Very
Secure FTP Daemon), but right now it only seems to work when I have the
line `force_local_data_ssl=NO`, when I think it should be `YES`.

Additionally, while the FTPS container can accept files when they are
sent, it doesn't share any information with the MySQL container, but
maybe this is as simple as mounting the Persistent Volume onto /var/ftp
in the ftps.yaml file (I think it really should be this simple).
In this way it should immediately be present in the mysql database -
however, I probably should take care to ensure that files are sent to
the correct location.