Testing looks good
I'd really like a second opinion on this of course, I'm sure there's
probably some things that are meant to be done that I'm completely
unaware of - but to me, this project looks pretty much finished.

Only change was to edit the title on the nginx homepage - wordpress
data and mysql data persisted after the pods were brought down - life is
good.